Move weighting into the search feature
